requirements.txt
asn1crypto==1.2.0
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
future==0.18.0
gevent==1.4.0
greenlet==0.4.15
hkdf==0.0.3
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
magic-wormhole==0.11.2
msgpack==0.6.2
nose==1.3.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyzmq==18.1.0
Random-Word==1.0.4
requests==2.22.0
service-identity==18.1.0
six==1.12.0
spake2==0.8
tqdm==4.36.1
Twisted==19.7.0
txaio==18.8.1
txtorcon==19.1.0
urllib3==1.25.6
zerorpc==0.6.3
zope.interface==4.6.0
